refactor(client): P10 — scheduled events, turn recorder, slash commands + final sweep (#143)
Final phase (RFC-001 §7 P10, metrics in RFC R6):
- scheduled_events.py — ScheduledEventHandlers: the action router
  (reminder/check/workflow/digest), workflow condition + on_failure
  semantics, structured-RBAC scheduled dispatch, digest collection,
  mention resolution, monitor/failure alerts. Verbatim host-based.
- turn_recorder.py — TurnRecorder: trajectory user-content recording,
  context-trace creation, turn persistence, lifecycle webhooks, and the
  reflection triggers (chat post-op + gated loop). Verbatim host-based
  (one spurious host prologue dropped where the method uses only class
  consts).
- slash_commands.py — register_commands(bot): the six commands, verbatim.
- Bot keeps delegates for everything external (scheduler.start is wired
  with bot methods; characterization + workflow tests drive bot._on_*).
- Declared migrations: _FakeClient and _LoopIterClient in
  test_trajectory_completeness construct TurnRecorder(self) (the
  established fake-host pattern; assertions untouched).
- Disclosed: long literals in the moved scheduled/slash code carry
  per-line noqa: E501 (verbatim-move; wrapping 20 f-strings invited
  transcription risk in the final pass). Orphaned client imports removed.

client.py: 1,789 -> 1,272 lines (campaign: 5,790 -> 1,272, -78%).
Characterization 127 green; full suite 6040 passed / 4 skipped; all three
new modules ruff-clean; RFC status flipped to CODE-COMPLETE with the
honest §10 metrics (two size targets missed deliberately — rationale in
R6).
