+++ docs/plans/client-decomposition-plan.md
@@ -2,7 +2,7 @@
 
 | | |
 |---|---|
-| **Status** | **APPROVED** (R1) — Odin approved as campaign plan of record, 2026-07-04. Implementation NOT started; each phase requires Aaron's per-phase approval, then branch → PR → Odin review |
+| **Status** | **CODE-COMPLETE** (2026-07-05) — P0–P10 + P5c implemented and Odin-reviewed on `refactor/client-decomposition`. §10 metrics reported in R6. Next: local soak-test deployment with all live data retained (R2), then Aaron's master-merge/pipeline decision |
 | **Author** | Claude (on behalf of Aaron / Calmingstorm) |
 | **Reviewer** | Odin |
 | **Date** | 2026-07-04 |
@@ -520,6 +520,7 @@ Legend: target modules are §6.1 names. "facade" = a delegating stub remains on 
   - getsource replacement policy retained unchanged (reviewer: "correct — keep it").
 - **Approval (2026-07-04)** — Odin re-reviewed R1 and **approved** the doc as the campaign plan of record: *"No remaining plan-level blockers. Next step is Phase 0 as its own PR, with the characterization harness doing the heavy lifting before anyone starts carving into the corpse."*
 - **R2 (2026-07-04)** — Execution model set by Aaron (direction approval): long-lived campaign branch `refactor/client-decomposition`; phase PRs merge into it with Odin review each; **no master merge, no release pipeline, no live deploy during the campaign**; post-P10 local soak test with all live data retained, then the pipeline decision. §7 execution-model paragraph updated; original per-phase release mapping retained inactive.
+- **R6 (2026-07-05)** — **Campaign code-complete.** P10 delivered `scheduled_events.py`, `turn_recorder.py`, `slash_commands.py` + final sweep. §10 metrics vs targets: **met** — one dispatch table (zero `elif tool_name ==` outside the registry), duplicate loop deleted, facade contract green throughout, external consumers (web/, health/, `__main__`, cogs) have ZERO diffs since baseline `081e231`, tests 6,040 ≥ baseline 5,905 (+127 characterization), suite runtime ~62s (≤ baseline), every phase Odin-reviewed and independently revertable. **Missed, deliberately** — `client.py` = 1,272 lines vs ≤~900 target (~450 lines are the one-line delegate/shim facade that bought zero external breakage; the true adapter is ~820) and `tool_loop.py` = 1,365 vs ≤~900 (it deliberately holds BOTH pipelines + LoopPolicy; re-splitting would recreate the dual-file sync hazard the campaign existed to kill). Long literals in moved scheduled/slash code carry per-line `noqa: E501` (verbatim-move disclosure) rather than wraps. Facade-delegate slimming, if ever desired, requires migrating `web/api.py` off private names first — explicitly out of scope (§3).
 - **R5 (2026-07-04)** — P7 delivered as a **move-only** extraction (reviewer-approved deviation): `_process_with_tools` lands verbatim in `ToolLoopRunner.run` (host-based, per the P5a pattern) and the promised decomposition into ~10 named methods **rides with P8**, where LoopPolicy parameterization forces exactly those seams — one restructuring pass over the loop's lines instead of two. CompletionClassifier extracted per plan (provider-backed llm client; facade delegates + class-attr alias retained).
 - **R4 (2026-07-04)** — P5b delivered 4 of 5 handler domains (scheduling, knowledge, channel-ops, media — 31 of the 46 handled tools). The **agents/tasks/loops domain is deferred to a P5c rider after P8**: its handlers close over the loop pipeline itself (`_handle_start_loop` → `_run_loop_iteration`, `delegate_task` → `llm_client`, spawn/collect → trajectory + lifecycle plumbing), which P7/P8 restructure — moving them first would mean moving them twice across the campaign's riskiest phases. Dispatch is unaffected (the table resolves them late on the host either way).
 - **R3 (2026-07-04)** — P5 executed via the declared escape hatch, **restructured**: P5a = the single `NativeToolDispatcher` table + BOTH chain swaps + `NativeToolEffects` + centralized skill-CRUD invalidation, with handler bodies remaining OdinBot methods registered by attribute name (late-bound at dispatch, preserving the historical `self._handle_X` lookup semantics and the test patch seam). P5b = mechanical relocation of handler bodies into `native_tools/` domain modules behind the same attribute names, dispatch untouched. Rationale: the semantic risk Odin flagged lives entirely in the swap; landing it first with the characterization suite green isolates that risk from the 1,200-line body move. The original "two commits in one PR" mandate maps onto the two PRs. Additional P5a decisions: RBAC checks stay in the callers verbatim (dispatcher does not re-check); executor fallthrough stays caller-side (chat needs the structured ToolResult, the loop wraps it); skill file delivery is a dispatch-time policy (`send` = chat, `stage` = loop, `export_skill` stages in both — matching baseline).
