+++ src/discord/client.py
@@ -8,15 +8,10 @@ import time
 from collections.abc import Callable
 
 import discord
-from discord import app_commands
 from discord.ext import commands
 
 from ..config.schema import Config
 from ..monitoring import InfraWatcher
-from .background_task import (
-    BackgroundTask, run_background_task, create_task_id, MAX_STEPS,
-)
-from ..agents.manager import AGENT_BLOCKED_TOOLS, filter_agent_tools
 from ..llm.secret_scrubber import scrub_output_secrets
 from ..odin_log import get_logger
 from ..tools import ToolResult, get_tool_definitions
@@ -32,6 +27,9 @@ from .delivery import SEND_MAX_RETRIES  # noqa: F401 — module re-export contra
 from .llm_gateway import LLMGateway
 from .native_tools import NativeToolDispatcher, register_native_handlers
 from .native_tools.agents_tasks import AgentTaskTools
+from .scheduled_events import ScheduledEventHandlers
+from .slash_commands import register_commands
+from .turn_recorder import TurnRecorder
 from .native_tools.channel_ops import ChannelOpsTools
 from .native_tools.media import MediaTools
 from .native_tools.knowledge import KnowledgeTools
@@ -348,6 +346,8 @@ class OdinBot(commands.Bot):
             get_llm_client=lambda: self.llm_client,
         )
         self._tool_loop_runner = ToolLoopRunner(self)
+        self._turn_recorder = TurnRecorder(self)
+        self._scheduled_events = ScheduledEventHandlers(self)
         self._agent_task_tools = AgentTaskTools(self)
         self._message_intake = MessageIntake(self)
         self._message_pipeline = MessagePipeline(self)
@@ -502,42 +502,14 @@ class OdinBot(commands.Bot):
         if cfg.discord.require_mention:
             log.info("Mention-only mode — will only respond when @mentioned")
 
+    # -- turn observability: bodies in turn_recorder.py (P10) --
+
     def _record_user_content(self, trajectory, content: str) -> None:
-        """Store the request on the trajectory turn — capped, secret-scrubbed,
-        with explicit truncation metadata. Config-gated; never raises."""
-        try:
-            obs = getattr(self.config, "observability", None)
-            if obs is not None and not obs.trajectory_user_content:
-                return
-            cap = getattr(obs, "max_user_content_chars", 4000) if obs else 4000
-            scrubbed = scrub_output_secrets(content)
-            if len(scrubbed) > cap:
-                trajectory.user_content_truncated = True
-                trajectory.user_content_original_chars = len(scrubbed)
-                scrubbed = scrubbed[:cap]
-            trajectory.user_content = scrubbed
-        except Exception:  # noqa: BLE001 — recording must never break the turn
-            log.debug("user_content recording failed (non-fatal)", exc_info=True)
+        self._turn_recorder._record_user_content(trajectory, content)
 
     def _new_context_trace(self):
-        """Create a context-trace collector when observability is enabled.
+        return self._turn_recorder._new_context_trace()
 
-        Returns None when disabled — every consumer treats None as
-        "record nothing", leaving the request path byte-identical.
-        """
-        try:
-            obs = getattr(self.config, "observability", None)
-            ct = getattr(obs, "context_trace", None)
-            if ct is None or not ct.enabled:
-                return None
-            from ..observability import ContextTraceCollector
-            return ContextTraceCollector(
-                memory_key_mode=ct.memory_key_mode,
-                include_segment_ids=ct.include_segment_ids,
-                max_trace_bytes=ct.max_trace_bytes,
-            )
-        except Exception:  # noqa: BLE001 — tracing must never block a turn
-            return None
 
     def _invalidate_prompt_caches(self) -> None:
         """Invalidate all prompt-related caches. Called on config/context reload."""
@@ -669,95 +641,9 @@ class OdinBot(commands.Bot):
         self._channel_state.track_recent_action(channel_id, entry)
 
     def _register_commands(self) -> None:
-        @self.tree.command(name="status", description="Show Odin bot status")
-        async def cmd_status(interaction: discord.Interaction) -> None:
-            if not self._is_allowed_user(interaction.user):
-                await interaction.response.send_message("Access denied.", ephemeral=True)
-                return
-            voice_status = ""
-            if self.voice_manager:
-                if self.voice_manager.is_connected:
-                    ch = self.voice_manager.current_channel
-                    voice_status = f"\nVoice: Connected to **{ch.name}**" if ch else "\nVoice: Connected"
-                else:
-                    voice_status = "\nVoice: Not connected"
-            provider_cfg = getattr(self.config, "llm_provider", None)
-            active = provider_cfg.active_provider if provider_cfg else "codex"
-            client = self.llm_client
-            if client:
-                model = getattr(client, "model", "unknown")
-                llm_status = f"LLM: **{active}** ({model})"
-            else:
-                llm_status = "LLM: not configured"
-            codex_configured = "yes" if self.codex_client else "no"
-            ollama_configured = "yes" if self.ollama_client else "no"
-            kimi_configured = "yes" if self.kimi_client else "no"
-            await interaction.response.send_message(
-                f"**Odin Status**\n"
-                f"{llm_status}\n"
-                f"Codex: {codex_configured} | Ollama: {ollama_configured} | Kimi: {kimi_configured}\n"
-                f"{voice_status}"
-            )
+        """Slash commands — owned by slash_commands.register_commands (P10)."""
+        register_commands(self)
 
-        @self.tree.command(name="reset", description="Reset conversation history")
-        async def cmd_reset(interaction: discord.Interaction) -> None:
-            if not self._is_allowed_user(interaction.user):
-                await interaction.response.send_message("Access denied.", ephemeral=True)
-                return
-            self.sessions.reset(str(interaction.channel_id))
-            await interaction.response.send_message("Conversation history cleared.")
-
-        @self.tree.command(name="reload", description="Reload context files")
-        async def cmd_reload(interaction: discord.Interaction) -> None:
-            if not self._is_allowed_user(interaction.user):
-                await interaction.response.send_message("Access denied.", ephemeral=True)
-                return
-            self.context_loader.reload()
-            self._invalidate_prompt_caches()
-            self._system_prompt = self._build_system_prompt()
-            await interaction.response.send_message("Context files reloaded.")
-
-        @self.tree.command(name="purge", description="Delete recent messages in this channel")
-        @app_commands.describe(count="Number of messages to delete (default 100, max 500)")
-        async def cmd_purge(interaction: discord.Interaction, count: int = 100) -> None:
-            if not self._is_allowed_user(interaction.user):
-                await interaction.response.send_message("Access denied.", ephemeral=True)
-                return
-            count = min(count, 500)
-            await interaction.response.defer(ephemeral=True)
-            deleted = await interaction.channel.purge(limit=count)
-            self.sessions.reset(str(interaction.channel_id))
-            await interaction.followup.send(
-                f"Deleted {len(deleted)} messages and reset conversation history.",
-                ephemeral=True,
-            )
-
-        @self.tree.command(name="usage", description="Show token usage details")
-        async def cmd_usage(interaction: discord.Interaction) -> None:
-            if not self._is_allowed_user(interaction.user):
-                await interaction.response.send_message("Access denied.", ephemeral=True)
-                return
-            await interaction.response.send_message(
-                "**Usage**\n"
-                "All backends are subscription-based (free).\n"
-                "Codex: ChatGPT subscription\n"
-                "Claude Code: Max subscription"
-            )
-
-
-        @self.tree.command(name="stop", description="Stop Odin's current task in this channel")
-        async def cmd_stop(interaction: discord.Interaction) -> None:
-            if not self._is_allowed_user(interaction.user):
-                await interaction.response.send_message("Access denied.", ephemeral=True)
-                return
-            channel_id = str(interaction.channel_id)
-            event = self._cancel_events.setdefault(channel_id, asyncio.Event())
-            active = self._channel_state.active_requests.get(channel_id)
-            if active:
-                event.set()
-                await interaction.response.send_message("Stopping current task...", ephemeral=True)
-            else:
-                await interaction.response.send_message("No active task in this channel.", ephemeral=True)
 
     def _is_cancelled(self, channel_id: str) -> bool:
         ev = self._cancel_events.get(channel_id)
@@ -801,139 +687,36 @@ class OdinBot(commands.Bot):
         )
 
 
-    # Successful operations below this tool count are routine — reflection
-    # is reserved for failures, corrections, explicit asks, and substantive work.
-    _REFLECT_MIN_TOOLS = 5
-    _REFLECT_CORRECTION_MARKERS = (
-        "remember this", "remember that", "that's wrong", "thats wrong",
-        "that is wrong", "not what i asked", "you should have", "incorrect,",
-        "no, ", "actually,",
-    )
+    async def _operational_reflection(
+        self, user_request: str, tools_used: list[str], response: str,
+        is_error: bool, user_id: str | None, tool_details: list[dict] | None = None,
+    ) -> None:
+        await self._turn_recorder._operational_reflection(
+            user_request, tools_used, response, is_error, user_id, tool_details=tool_details,
+        )
 
     def _should_reflect_on_operation(
-        self, user_request: str, tools_used: list[str],
-        is_error: bool, tool_details: list[dict],
+        self, user_request: str, tools_used: list[str], is_error: bool, tool_details: list[dict],
     ) -> bool:
-        """Reflection triggers: failure, mid-operation tool errors (recovery),
-        user corrections, explicit remember-this, or substantive operations.
-        Routine successes (ls/git-status class) skip reflection entirely."""
-        if is_error:
-            return True
-        if any(d.get("error") for d in tool_details):
-            return True
-        req = user_request.lower()
-        if any(marker in req for marker in self._REFLECT_CORRECTION_MARKERS):
-            return True
-        return len(tools_used) >= self._REFLECT_MIN_TOOLS
+        return self._turn_recorder._should_reflect_on_operation(
+            user_request, tools_used, is_error, tool_details,
+        )
 
-    def _maybe_loop_reflect(
-        self, *, loop_id: str, prompt: str, outcome: str, is_error: bool,
-        failure_class: str, error_text: str, tool_details: list[dict],
-        user_id: str | None,
-    ) -> None:
-        """Gated reflection for loop iterations (fire-and-forget)."""
-        try:
-            if not getattr(self.config.learning, "loop_reflection_enabled", True):
-                return
-            if not hasattr(self, "reflector") or not tool_details:
-                return
-            if is_error or any(d.get("error") for d in tool_details):
-                effective_error = error_text or next(
-                    (d["result"] for d in tool_details if d.get("error")), "",
-                )
-                if not failure_class:
-                    from ..observability.failure_classes import classify_failure
-                    failure_class = classify_failure(effective_error)["class"]
-                should, reason = self._loop_reflection_gate.evaluate(
-                    loop_id, is_error=True,
-                    failure_class=failure_class, error_text=effective_error,
-                )
-            else:
-                should, reason = self._loop_reflection_gate.evaluate(
-                    loop_id, is_error=False,
-                )
-            if not should:
-                log.debug("Loop reflection suppressed (%s) for %s", reason, loop_id)
-                return
-            log.info("Loop reflection triggered (%s) for %s", reason, loop_id)
-            fire_and_forget(self.reflector.reflect_on_operation(
-                user_request=f"[autonomous loop {loop_id}] {prompt[:300]}",
-                tools_used=[d["tool"] for d in tool_details][:20],
-                tool_details=tool_details,
-                final_response=outcome,
-                is_error=is_error,
-                user_id=user_id,
-            ), name="loop_reflection")
-        except Exception:  # noqa: BLE001 — reflection must never break a loop
-            log.debug("Loop reflection wiring failed (non-fatal)", exc_info=True)
-
-    async def _operational_reflection(
-        self, user_request: str, tools_used: list[str],
-        response: str, is_error: bool, user_id: str | None,
-        tool_details: list[dict] | None = None,
-    ) -> None:
-        """Fire-and-forget post-operation reflection — selective, with real
-        tool inputs/results from the operation instead of bare tool names."""
-        try:
-            if not tool_details:
-                tool_details = [{"tool": t} for t in tools_used[:20]]
-            if not self._should_reflect_on_operation(
-                user_request, tools_used, is_error, tool_details,
-            ):
-                log.debug(
-                    "Skipping reflection for routine operation (%d tools, no errors)",
-                    len(tools_used),
-                )
-                return
-            await self.reflector.reflect_on_operation(
-                user_request=user_request,
-                tools_used=tools_used,
-                tool_details=tool_details,
-                final_response=response,
-                is_error=is_error,
-                user_id=user_id,
-            )
-        except Exception as e:
-            log.debug("Operational reflection failed (non-fatal): %s", e)
+    def _maybe_loop_reflect(self, **kwargs) -> None:
+        self._turn_recorder._maybe_loop_reflect(**kwargs)
 
     async def _save_turn_trajectory(
         self, trajectory, *, error: str = "", final_response: str = "",
         tools_used: list[str] | None = None, trace=None,
     ) -> None:
-        """Persist the turn trajectory as JSONL. Non-fatal on error."""
-        if self.trajectory_saver is None:
-            return
-        try:
-            if trace is not None:
-                trajectory.context_trace = trace.finalize()
-            from datetime import datetime, timezone
-            trajectory.timestamp = datetime.now(timezone.utc).isoformat()
-            if error:
-                trajectory.is_error = True
-                trajectory.final_response = error
-            elif final_response:
-                trajectory.final_response = final_response
-            if tools_used is not None:
-                trajectory.tools_used = list(tools_used)
-            # Aggregate token counts from iterations
-            trajectory.total_input_tokens = sum(it.input_tokens for it in trajectory.iterations)
-            trajectory.total_output_tokens = sum(it.output_tokens for it in trajectory.iterations)
-            await self.trajectory_saver.save(trajectory)
-        except Exception:
-            log.exception("TrajectorySaver.save failed (non-fatal)")
+        await self._turn_recorder._save_turn_trajectory(
+            trajectory, error=error, final_response=final_response,
+            tools_used=tools_used, trace=trace,
+        )
 
     async def _emit_lifecycle_event(self, event_type: str, payload: dict) -> None:
-        """Emit a lifecycle event to registered outbound webhooks (no-op if disabled)."""
-        if self.outbound_webhook_dispatcher is None:
-            return
-        try:
-            from ..notifications.outbound_webhooks import build_event_payload
-            full_payload = build_event_payload(event_type=event_type, data=payload)
-            await self.outbound_webhook_dispatcher.dispatch_fire_and_forget(
-                event_type=event_type, payload=full_payload,
-            )
-        except Exception:
-            log.exception("Outbound webhook dispatch failed (non-fatal)")
+        await self._turn_recorder._emit_lifecycle_event(event_type, payload)
+
 
     async def setup_hook(self) -> None:
         """Called once before connecting to the gateway.
@@ -1335,114 +1118,20 @@ class OdinBot(commands.Bot):
         return await self._knowledge_tools._handle_search_audit(inp)
 
 
+    # -- scheduler/digest/monitor callbacks: bodies in scheduled_events.py (P10) --
+
     async def _on_scheduled_digest(self, schedule: dict) -> None:
-        """Run the daily infrastructure digest and post results."""
-        channel_id = schedule.get("channel_id")
-        if not channel_id:
-            log.warning("Digest has no channel_id: %s", schedule["id"])
-            return
-
-        channel = self.get_channel(int(channel_id))
-        if not channel:
-            log.warning("Digest channel %s not found", channel_id)
-            return
-
-        log.info("Running daily digest for channel %s", channel_id)
-        try:
-            raw = await self._format_digest_raw()
-        except Exception as e:
-            log.error("Digest data collection failed: %s", e)
-            await channel.send(scrub_response_secrets(f"**Daily Infrastructure Digest**\n\nFailed to collect data: {e}"))
-            return
-
-        # Summarize the digest — prefer Codex (free), fall back to raw truncation
-        digest_messages = [{"role": "user", "content": f"Summarize this infrastructure status report concisely. Highlight any issues, warnings, or anomalies. If everything looks healthy, say so briefly.\n\n{raw}"}]
-        digest_system = "You are a concise infrastructure report summarizer. Output a short summary with key findings."
-        try:
-            if self.llm_client:
-                summary = await self.llm_client.chat(
-                    messages=digest_messages, system=digest_system, max_tokens=500,
-                )
-            else:
-                log.warning("No Codex client for digest summary, using raw")
-                summary = raw[:3000]
-        except Exception as e:
-            log.warning("Digest summary failed, using raw: %s", e)
-            summary = raw[:3000]
-
-        await channel.send(scrub_response_secrets(f"**Daily Infrastructure Digest**\n\n{summary}"))
-
-        # Audit log the digest
-        await self.audit.log_execution(
-            user_id="system",
-            user_name="scheduler",
-            channel_id=channel_id,
-            tool_name="digest",
-            tool_input={"schedule_id": schedule.get("id")},
-            approved=True,
-            result_summary=summary,
-            execution_time_ms=0,
-        )
+        await self._scheduled_events._on_scheduled_digest(schedule)
 
     async def _format_digest_raw(self) -> str:
-        """Collect raw infrastructure data for the digest."""
-        tasks = []
-        labels = []
-
-        # Disk + memory checks on all hosts via run_command
-        for host_alias in self.config.tools.hosts:
-            tasks.append(self.tool_executor.execute(
-                "run_command", {"host": host_alias, "command": "df -h --exclude-type=tmpfs --exclude-type=devtmpfs"},
-            ))
-            labels.append(f"Disk ({host_alias})")
-            tasks.append(self.tool_executor.execute(
-                "run_command", {"host": host_alias, "command": "free -h"},
-            ))
-            labels.append(f"Memory ({host_alias})")
-
-        results = await asyncio.gather(*tasks, return_exceptions=True)
-
-        sections = []
-        for label, result in zip(labels, results):
-            if isinstance(result, Exception):
-                sections.append(f"### {label}\nERROR: {result}")
-            else:
-                sections.append(f"### {label}\n{str(result)[:800]}")
-
-        return "\n\n".join(sections)
+        return await self._scheduled_events._format_digest_raw()
 
     def _resolve_mentions(self, text: str) -> str:
-        """Replace @username with proper Discord <@ID> mentions."""
-        def _replace(match: re.Match) -> str:
-            name = match.group(1).lower()
-            for guild in self.guilds:
-                for member in guild.members:
-                    if member.name.lower() == name or (member.nick and member.nick.lower() == name):
-                        return f"<@{member.id}>"
-            return match.group(0)  # leave unchanged if not found
-        return re.sub(r"@(\w+)", _replace, text)
+        return self._scheduled_events._resolve_mentions(text)
 
     async def _on_monitor_alert(self, message: str) -> None:
-        """Callback fired by the infrastructure watcher when a threshold is crossed."""
-        channel_id = self.config.monitoring.alert_channel_id
-        if not channel_id:
-            # Fall back to first configured channel
-            if self.config.discord.channels:
-                channel_id = self.config.discord.channels[0]
-            else:
-                log.warning("Monitor alert has no channel to send to: %s", message[:100])
-                return
+        await self._scheduled_events._on_monitor_alert(message)
 
-        channel = self.get_channel(int(channel_id))
-        if not channel:
-            log.warning("Monitor alert channel %s not found", channel_id)
-            return
-
-        try:
-            await channel.send(scrub_response_secrets(message))
-            log.info("Sent monitor alert to channel %s", channel_id)
-        except Exception as e:
-            log.error("Failed to send monitor alert: %s", e)
 
     # -- agents/tasks/loops handlers: bodies in native_tools/agents_tasks.py (P5c) --
 
@@ -1546,233 +1235,22 @@ class OdinBot(commands.Bot):
 
 
     async def _execute_scheduled_tool(
-        self,
-        tool_name: str,
-        tool_input: dict,
-        channel: discord.abc.Messageable,
-        requester_id: str | None,
-        requester_name: str = "scheduler",
+        self, tool_name: str, tool_input: dict, channel: discord.abc.Messageable,
+        requester_id: str | None, requester_name: str = "scheduler",
     ) -> ToolResult:
-        """Unified dispatch for scheduled task tool execution.
+        return await self._scheduled_events._execute_scheduled_tool(
+            tool_name, tool_input, channel, requester_id, requester_name,
+        )
 
-        Routes through the same client-level dispatch as live messages and
-        autonomous loops, so scheduled tasks have full tool parity.
-        """
-        # Pre-check RBAC so we can return a structured failure — the dispatch
-        # also checks, but returns a plain denial string we'd have to guess at.
-        if requester_id:
-            denial = self.tool_executor.check_permission(tool_name, requester_id)
-            if isinstance(denial, str) and denial:
-                return ToolResult(output=denial, ok=False, error="permission_denied", tool_name=tool_name)
-
-        msg_proxy = _LoopMessageProxy(channel, requester_id or "0", requester_name)
-        try:
-            result = await self._dispatch_loop_tool_inner(
-                tool_name, tool_input, msg_proxy, requester_id or "",
-            )
-        except Exception as e:
-            return ToolResult(output=f"Error executing {tool_name}: {e}", ok=False, error="execution_error", tool_name=tool_name)
-
-        if isinstance(result, ToolResult):
-            return result
-        return ToolResult(output=str(result), ok=True, tool_name=tool_name)
-
-    async def _run_scheduled_workflow(
-        self, channel: discord.abc.Messageable, schedule: dict,
-    ) -> bool:
-        """Execute a multi-step workflow from a scheduled task.
-
-        Returns True if all steps succeeded, False if any step failed.
-        """
-        steps = schedule.get("steps", [])
-        desc = schedule.get("description", "Workflow")
-        results: list[str] = []
-        prev_output = ""
-        workflow_ok = True
-        # Scheduled work runs under the identity of whoever created the schedule, so
-        # host-access scoping / tier limits apply (None = unrestricted system task).
-        req_id = schedule.get("requester_id") or None
-
-        for i, step in enumerate(steps):
-            tool_name = step["tool_name"]
-            tool_input = step.get("tool_input", {})
-            condition = step.get("condition")
-            step_desc = step.get("description", tool_name)
-
-            # Evaluate condition against previous step's output
-            if condition and prev_output:
-                if condition.startswith("!"):
-                    # Negated condition: skip if substring IS present
-                    if condition[1:].lower() in prev_output.lower():
-                        results.append(f"**Step {i+1}** (`{step_desc}`): skipped (condition `{condition}` met)")
-                        continue
-                else:
-                    # Normal condition: skip if substring is NOT present
-                    if condition.lower() not in prev_output.lower():
-                        results.append(f"**Step {i+1}** (`{step_desc}`): skipped (condition `{condition}` not met)")
-                        continue
-
-            try:
-                req_name = schedule.get("requester") or schedule.get("created_by") or "scheduler"
-                # Signal to spawn_agent handler that this is a scheduled context
-                if tool_name == "spawn_agent":
-                    tool_input = {**tool_input, "_scheduled": True}
-
-                result = await self._execute_scheduled_tool(
-                    tool_name, tool_input, channel, req_id, req_name,
-                )
-
-                # Auto-collect agent results for spawn_agent steps in scheduled workflows.
-                # Extract agent_id from spawn confirmation and wait for completion so the
-                # workflow reports what the agent did, not just that it was spawned.
-                render_markdown = False
-                if tool_name == "spawn_agent" and isinstance(result, ToolResult) and result.ok:
-                    result_str = str(result)
-                    id_match = re.search(r"\(ID:\s*`([^`]+)`\)", result_str)
-                    if id_match:
-                        agent_id = id_match.group(1)
-                        timeout = float(step.get("timeout", 3660))
-                        agent_text, agent_data = await self._collect_agent_result(
-                            agent_id, timeout=min(timeout, 3660),
-                        )
-                        agent_ok = agent_data["status"] == "completed"
-                        result = ToolResult(output=agent_text, ok=agent_ok, tool_name="spawn_agent")
-                        if agent_ok and agent_data["empty_result"]:
-                            result = ToolResult(
-                                output=agent_text + "\n\n⚠️ Agent completed but produced no output.",
-                                ok=True, tool_name="spawn_agent",
-                            )
-                        render_markdown = True
-
-                prev_output = str(result)
-
-                if isinstance(result, ToolResult) and not result.ok:
-                    results.append(f"**Step {i+1}** (`{step_desc}`): FAILED\n```\n{str(result)[:1600]}\n```")
-                    on_failure = step.get("on_failure", "abort")
-                    if on_failure == "abort":
-                        workflow_ok = False
-                        results.append("Workflow aborted due to step failure.")
-                        break
-                elif render_markdown:
-                    results.append(f"**Step {i+1}** (`{step_desc}`): OK\n\n{str(result)[:1600]}")
-                else:
-                    results.append(f"**Step {i+1}** (`{step_desc}`): OK\n```\n{str(result)[:1600]}\n```")
-            except Exception as e:
-                results.append(f"**Step {i+1}** (`{step_desc}`): FAILED — {e}")
-                on_failure = step.get("on_failure", "abort")
-                if on_failure == "abort":
-                    workflow_ok = False
-                    results.append("Workflow aborted due to step failure.")
-                    break
-
-        summary = "\n".join(results)
-        text = f"**Workflow: {desc}**\n{summary}"
-        if len(text) > 1900:
-            text = text[:1900] + "\n... (truncated)"
-
-        try:
-            await channel.send(scrub_response_secrets(text))
-        except Exception as e:
-            log.error("Failed to post workflow results: %s", e)
-
-        return workflow_ok
+    async def _run_scheduled_workflow(self, channel: discord.abc.Messageable, schedule: dict) -> bool:
+        return await self._scheduled_events._run_scheduled_workflow(channel, schedule)
 
     async def _on_schedule_failure(self, schedule: dict, consecutive: int) -> None:
-        """Alert callback fired when a schedule crosses the consecutive-failure
-        threshold. Previously never wired, so the alerting path was dead."""
-        channel_id = schedule.get("channel_id")
-        last_error = schedule.get("last_error", "unknown error")
-        text = (
-            f"⚠️ **Scheduled task failing:** {schedule.get('description', schedule.get('id', '?'))}\n"
-            f"{consecutive} consecutive failures. Last error:\n"
-            f"```\n{str(last_error)[:1000]}\n```"
-        )
-        try:
-            channel = self.get_channel(int(channel_id)) if channel_id else None
-            if channel:
-                await channel.send(scrub_response_secrets(text))
-            else:
-                log.warning(
-                    "Schedule %s failed %d times but channel %s is unavailable for alert",
-                    schedule.get("id"), consecutive, channel_id,
-                )
-        except Exception as e:
-            log.warning("Failed to send schedule failure alert for %s: %s", schedule.get("id"), e)
+        await self._scheduled_events._on_schedule_failure(schedule, consecutive)
 
     async def _on_scheduled_task(self, schedule: dict) -> None:
-        """Callback fired by the scheduler when a task is due."""
-        try:
-            await self.audit.log_event(
-                event_type="schedule_execution",
-                action=schedule.get("action", "unknown"),
-                actor="scheduler",
-                detail=f"Schedule {schedule.get('id', '?')}: {schedule.get('description', '')[:100]}",
-                channel_id=schedule.get("channel_id", ""),
-                metadata={"schedule_id": schedule.get("id"), "action": schedule.get("action")},
-            )
-        except Exception:
-            pass
-        channel_id = schedule.get("channel_id")
-        if not channel_id:
-            log.warning("Scheduled task has no channel_id: %s", schedule["id"])
-            return
+        await self._scheduled_events._on_scheduled_task(schedule)
 
-        channel = self.get_channel(int(channel_id))
-        if not channel:
-            log.warning("Scheduled task channel %s not found", channel_id)
-            return
-
-        if schedule["action"] == "digest":
-            await self._on_scheduled_digest(schedule)
-            return
-
-        if schedule["action"] == "reminder":
-            msg = schedule.get("message", schedule["description"])
-            # Resolve @username mentions to proper Discord <@ID> mentions
-            msg = self._resolve_mentions(msg)
-            try:
-                await channel.send(f"**Scheduled reminder:** {msg}")
-            except Exception as e:
-                log.warning("Failed to send scheduled reminder: %s", e)
-
-        elif schedule["action"] == "check":
-            tool_name = schedule.get("tool_name")
-            tool_input = schedule.get("tool_input", {})
-            req_id = schedule.get("requester_id") or None
-            req_name = schedule.get("requester") or schedule.get("created_by") or "scheduler"
-            try:
-                result = await self._execute_scheduled_tool(
-                    tool_name, tool_input, channel, req_id, req_name,
-                )
-                if isinstance(result, ToolResult) and not result.ok:
-                    text = f"**Scheduled check failed:** {schedule['description']}\n```\n{str(result)[:1800]}\n```"
-                    try:
-                        await channel.send(scrub_response_secrets(text))
-                    except Exception:
-                        pass
-                    raise RuntimeError(f"Scheduled check failed: {str(result)[:200]}")
-                else:
-                    text = f"**Scheduled: {schedule['description']}**\n```\n{str(result)[:1800]}\n```"
-                    await channel.send(scrub_response_secrets(text))
-            except RuntimeError:
-                raise
-            except Exception as e:
-                log.error("Scheduled task error: %s", e, exc_info=True)
-                try:
-                    await channel.send(
-                        scrub_response_secrets(f"**Scheduled task failed:** {schedule['description']}\nError: {e}")
-                    )
-                except Exception:
-                    pass
-                raise
-
-        elif schedule["action"] == "workflow":
-            ok = await self._run_scheduled_workflow(channel, schedule)
-            if not ok:
-                raise RuntimeError(f"Scheduled workflow failed: {schedule.get('description', '')[:200]}")
-
-        else:
-            log.warning("Unknown scheduled action type: %s (schedule %s)", schedule["action"], schedule.get("id"))
 
     async def _send_with_retry(
         self,
